requirements.txt
sentence-transformers>=2.6.0
psycopg2-binary>=2.9.9
numpy>=1.26.0
pdfplumber>=0.11.0